Added a few improvements for AGP printing
The code will now look specifically for American Gut Project samples and
will not print anything that's not under this classification in the
TITLE column. If one is found like this, the program will warn you about
this.

A few alerts have been added as well as a check box to do a test run for
the code to only generate the global image and one sample.